Unit and Functionnal tests

+++ src/Controller/TaskController.php
@@ -3,9 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 class TaskController extends AbstractController
@@ -13,9 +15,17 @@ class TaskController extends AbstractController
 
     public function listAction()
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+        $user = $this->getUser();
+
+        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findNotDone(), 'user' => $user]);
     }
 
+    public function listIsDoneAction()
+    {
+        $user = $this->getUser();
+
+        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findIsDone(), 'user' => $user]);
+    }
 
     public function createAction(Request $request)
     {
@@ -26,6 +36,12 @@ class TaskController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
+
+            $user = $this->getUser();
+            $userid = $user->getId();
+            $repository = $this->getDoctrine()->getRepository(User::class);
+            $user = $repository->findOneBy(['id' => $userid]);
+            $task->setUserCreate($user);
 
             $em->persist($task);
             $em->flush();
@@ -41,16 +57,29 @@ class TaskController extends AbstractController
 
     public function editAction(Task $task, Request $request)
     {
-        $form = $this->createForm(TaskType::class, $task);
+        $taskid = $task->getId();
+        $user = $this->getUser();
+        $userid = $user->getId();
+        $roles = $user->getRoles();
+        $repository = $this->getDoctrine()->getRepository(Task::class);
+        $taskvalid = $repository->findOneBy(['id' => $taskid, 'userCreate' => $userid]);
 
-        $form->handleRequest($request);
+        if (!empty($taskvalid) or $roles[0] == 'ROLE_ADMIN') {
+          $form = $this->createForm(TaskType::class, $task);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+          $form->handleRequest($request);
 
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('task_list');
+                $this->addFlash('success', 'La tâche a bien été modifiée.');
+
+                return $this->redirectToRoute('task_list');
+            }
+        }
+        else {
+          $this->addFlash("error", "Cette tâche ne vous appartient pas");
+          return $this->redirectToRoute('task_list');
         }
 
         return $this->render('task/edit.html.twig', [
@@ -65,20 +94,42 @@ class TaskController extends AbstractController
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
 
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        $value = $task->getIsDone();
 
-        return $this->redirectToRoute('task_list');
+        if ($value == '0') {
+          $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme non faite.', $task->getTitle()));
+
+          return $this->redirectToRoute('task_list');
+        }
+        elseif ($value == '1') {
+          $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+
+          return $this->redirectToRoute('task_list_done');
+        }
     }
 
 
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+      $taskid = $task->getId();
+      $user = $this->getUser();
+      $userid = $user->getId();
+      $roles = $user->getRoles();
+      $repository = $this->getDoctrine()->getRepository(Task::class);
+      $taskvalid = $repository->findOneBy(['id' => $taskid, 'userCreate' => $userid]);
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+      if (!empty($taskvalid) or $roles[0] == 'ROLE_ADMIN') {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+            return $this->redirectToRoute('task_list');
+        }
+      else {
+        $this->addFlash("error", "Cette tâche ne vous appartient pas");
         return $this->redirectToRoute('task_list');
+      }
     }
 }
